change id type to Integer

+++ backend/booking-service/src/main/java/com/hotel/booking/guest/GoscRepository.java
@@ -2,5 +2,5 @@ package com.hotel.booking.guest;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GoscRepository extends JpaRepository<Gosc, Long> {
+public interface GoscRepository extends JpaRepository<Gosc, Integer> {
 }
