+++ backend/catalog-service/src/main/java/com/hotel/catalog/hotel/HotelRepository.java
@@ -2,4 +2,4 @@ package com.hotel.catalog.hotel;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HotelRepository extends JpaRepository<Hotel, Long> {}
+public interface HotelRepository extends JpaRepository<Hotel, Integer> {}
