+++ backend/catalog-service/src/main/java/com/hotel/catalog/room/PokojRepository.java
@@ -2,5 +2,5 @@ package com.hotel.catalog.room;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PokojRepository extends JpaRepository<Pokoj, Long> {
+public interface PokojRepository extends JpaRepository<Pokoj, Integer> {
 }
